fix: rename UserService.js to userService.js (case-only change)
